Ran for varying sizes. Problems at the borders
The difference plateaus for larger tesseroids. This is probably because
the difference between the tesseroid geometry and the prism model gets
larger as well. Increasing discretization a reasonable amount did not
help. The difference concentrates on the border of the tesseroid.

Possible fix: Calculate only on top of the tesseroid. This would
minimize the effects of the prism approximation.