Updated shell comparison with new division scheme
This reflects the 'divide' branch of Tesseroids. The ratios are higher
but the software runs faster because of smarter division.
Potential still has error that varies with size. I think it's normal
because other vary as well, only less. Notebook also now contains more
provenance about the version of Tesseroids.