Calculating on the pole instead
There are no more singularities so I can calculate on the pole now. That
makes the argument stronger because the pole is weird on the
tesseroids. And if it works there...
Need to run the notebook on the whole data set as well.
This reflects a change in Tesseroids on the way I was calculating the
size of the tesseroid. It was wrong. I have implemented a true
great-circle computation to calculate the size so the division should
work on the pole as well.